Application structure improved
