+++ alayatodo/__init__.py
@@ -12,6 +12,16 @@ PASSWORD = 'default'
 app = Flask(__name__)
 app.config.from_object(__name__)
 
+# blueprints
+from alayatodo._auth import bp as _auth_bp
+app.register_blueprint(_auth_bp)
+
+from alayatodo._todo import bp as _todo_bp
+app.register_blueprint(_todo_bp)
+
+from alayatodo._main import bp as _main_bp
+app.register_blueprint(_main_bp)
+
 
 def connect_db():
     conn = sqlite3.connect(app.config['DATABASE'])
@@ -29,6 +39,3 @@ def teardown_request(exception):
     db = getattr(g, 'db', None)
     if db is not None:
         db.close()
-
-
-import alayatodo.views
