+++ alayatodo/_auth/__init__.py
@@ -1,0 +1,5 @@
+from flask import Blueprint
+
+bp = Blueprint('_auth', __name__)
+
+from alayatodo._auth import routes
