+++ alayatodo/_auth/routes.py
@@ -1,0 +1,28 @@
+from alayatodo._auth import bp
+from flask import g, render_template, request, session, redirect
+
+
+@bp.route('/login', methods=['GET'])
+def login():
+    return render_template('login.html')
+
+@bp.route('/login', methods=['POST'])
+def login_POST():
+    username = request.form.get('username')
+    password = request.form.get('password')
+
+    sql = "SELECT * FROM users WHERE username = '%s' AND password = '%s'";
+    cur = g.db.execute(sql % (username, password))
+    user = cur.fetchone()
+    if user:
+        session['user'] = dict(user)
+        session['logged_in'] = True
+        return redirect('/todo')
+
+    return redirect('/login')
+
+@bp.route('/logout')
+def logout():
+    session.pop('logged_in', None)
+    session.pop('user', None)
+    return redirect('/')
