+++ alayatodo/_main/__init__.py
@@ -1,0 +1,5 @@
+from flask import Blueprint
+
+bp = Blueprint('_main', __name__)
+
+from alayatodo._main import routes
