+++ alayatodo/_todo/__init__.py
@@ -1,0 +1,5 @@
+from flask import Blueprint
+
+bp = Blueprint('_todo', __name__)
+
+from alayatodo._todo import routes
