+++ alayatodo/templates/todo.html
@@ -10,7 +10,7 @@
             <td>{{ todo.user_id }}</td>
             <td>{{ todo.description }}</td>
             <td>
-                <form method="post" action="{{ url_for('todo_delete', id=todo.id) }}">
+                <form method="post" action="{{ url_for('_todo.todo_delete', id=todo.id) }}">
                     <button type="submit" class="btn btn-xs btn-danger"><span class="glyphicon glyphicon-remove glyphicon-white"></span></button>
                 </form>
             </td>
